crediblesource-server/requirements.txt
flask
ariadne
flask-sqlalchemy
flask-cors
psycopg2-binary